Unify CLI: Single ampere command with hybrid animated dashboard
Merged start, run, and demo into one ampere command with HybridDashboardRenderer
always enabled. Removed 7 TUI command subcommands (StartCommand, RunCommand,
DemoCommand, WatchCommand, DashboardCommand, InteractiveCommand, HelpCommand)
in favor of a single root command with integrated flags (--goal, --issues,
--issue, --arc, --list-arcs, --auto-work, --use-arc-phases). Utility
subcommands (thread, status, outcomes, knowledge, trace, task, issues, work,
respond, test) are kept. All tests pass.
