+++ .conductor/run.sh
@@ -5,6 +5,6 @@ echo "=== JVM Unit Tests ==="
 ./gradlew jvmTest
 
 echo "=== Build ==="
-./gradlew build -x test
+./gradlew assemble
 
 echo "=== All checks passed ==="
